feat(placements): one queue page for stickers and remixes
Stickers and remixes were two pages doing the same job — same `Placement`
table, same pending/expiry mechanics, same received/placed split, and two
separate things a creator had to remember to go and look at. Neither was
reachable from anywhere but account settings.

Now one route. `/user/placements` carries a segmented control for the two
surfaces with each one's received count on it, so "is there anything for me"
is answered in one look rather than by visiting two pages. Received/placed
sit underneath, unchanged: each queue keeps its own tabs, queries and row
rendering. Unifying the shell was the point; unifying the rows would have
been a rewrite of two working surfaces.

- The two old pages become their queue components verbatim, and the routes
  become redirects that carry `tab` through. A notification sent last week is
  a link someone clicks next month.
- The user menu entry is now "Placements" and counts both surfaces, since it
  points at both. NEW badge then the count, both immediately after the label.
- Remix pending notifications land on the queue too. They were deliberately
  left alone last time — one surface's fix should not silently become
  another's — but that stopped being a widening the moment the two queues
  became one page.
- The counts are one grouped query over `Placement` rather than one per
  surface: same index, one round trip, on a request that already runs once
  per session for every signed-in user.
- All three count keys are in the mark-all-read guard set. Wiping only the
  per-surface pair would empty the control's badges while the total survived
  — a half-fix that looks whole.

No avatar dot indicator: notifications already cover it.

The new browser test's first version mocked `next/router` itself, which the
shared harness already mocks — a setup-file mock wins, so mine did nothing,
and 5 of its 7 tests passed anyway because the fallback surface is the same
one the default renders. Only `?type=remix` could see it. It now drives the
harness's router, and the surface assertion was confirmed red against a
pinned surface.

Verification: typecheck 0 errors, 19,748 unit tests pass, browser suites
14/14, lint clean on every touched file.

Session: 60bc2834-9607-4106-9490-18bad86740ce
