+++ src/components/AppLayout/AppHeader/UserMenu.tsx
@@ -250,12 +250,15 @@ function UserMenuItems({ items }: { items: UserMenuItem[] }) {
               <item.icon stroke={1.5} color={item.color} />
               <span className="text-sm leading-none">{item.label}</span>
               {item.currency && <CurrencyIcon currency={Currency.BUZZ} size={16} />}
-              <QueueCountBadge count={item.badge ?? 0} max={99} ml={4} />
+              {/* NEW first, then the count, both immediately after the label
+                  rather than pushed to the right edge — the two badges read as
+                  one group belonging to the entry. */}
               {!!item.newUntil && Date.now() < item.newUntil.getTime() && (
                 <Badge color="green.8" variant="filled" size="sm" ml={4}>
                   NEW
                 </Badge>
               )}
+              <QueueCountBadge count={item.badge ?? 0} max={99} ml={4} />
             </>
           );
           const linkOrButton = item.href ? (
